+++ cpp/order/HexTransOrderParameter.cc
@@ -12,17 +12,12 @@ void HexTransOrderParameter<T>::computeGeneral(Func func, const freud::locality:
     m_box = points->getBox();
     unsigned int Np = points->getNPoints();
 
-    // Reallocate the output array if it is not the right size
-    if (Np != m_Np)
-    {
-        m_psi_array = std::shared_ptr<std::complex<float>>(new std::complex<float>[Np],
-                                                      std::default_delete<std::complex<float>[]>());
-    }
+    m_psi_array.prepare(Np);
 
     freud::locality::loopOverNeighborsIterator(points, points->getPoints(), Np, qargs, nlist, 
     [=] (size_t i, std::shared_ptr<freud::locality::NeighborPerPointIterator> ppiter)
     {
-        m_psi_array.get()[i] = 0;
+        m_psi_array[i] = 0;
         vec3<float> ref = (*points)[i];
 
         for(freud::locality::NeighborBond nb = ppiter->next(); !ppiter->end(); nb = ppiter->next())
@@ -32,10 +27,10 @@ void HexTransOrderParameter<T>::computeGeneral(Func func, const freud::locality:
 
             // Compute psi for neighboring particle
             // (only constructed for 2d)
-            m_psi_array.get()[i] += func(delta);
+            m_psi_array[i] += func(delta);
         }
 
-        m_psi_array.get()[i] /= std::complex<float>(m_k);
+        m_psi_array[i] /= std::complex<float>(m_k);
     });
     // Save the last computed number of particles
     m_Np = Np;